refactor: moved e2e create commands
