+++ e2e/test/commands/create/dart_cli/dart_cli_test.dart
@@ -1,17 +1,14 @@
-@Tags(['e2e'])
-library flame_game_test;
-
 import 'package:mason/mason.dart';
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
 import 'package:universal_io/io.dart';
 
-import '../../../../../helpers/helpers.dart';
+import '../../../../helpers/helpers.dart';
 
 void main() {
   test(
-    'create flame_game',
-    timeout: const Timeout(Duration(minutes: 5)),
+    'create dart_cli',
+    timeout: const Timeout(Duration(minutes: 2)),
     withRunner((commandRunner, logger, updater, logs) async {
       final tempDirectory = Directory.systemTemp.createTempSync();
       addTearDown(() => tempDirectory.deleteSync(recursive: true));
@@ -19,17 +16,22 @@ void main() {
       final result = await commandRunner.run(
         [
           'create',
-          'flame_game',
-          'very_good_flame_game',
+          'dart_cli',
+          'very_good_dart_cli',
           '-o',
           tempDirectory.path,
         ],
       );
       expect(result, equals(ExitCode.success.code));
 
-      final workingDirectory = path.join(
-        tempDirectory.path,
-        'very_good_flame_game',
+      final workingDirectory =
+          path.join(tempDirectory.path, 'very_good_dart_cli');
+
+      // add coverage to collect coverage on dart test
+      await expectSuccessfulProcessResult(
+        'dart',
+        ['pub', 'add', 'coverage:1.2.0'],
+        workingDirectory: workingDirectory,
       );
 
       await expectSuccessfulProcessResult(
@@ -46,18 +48,34 @@ void main() {
       expect(analyzeResult.stdout, contains('No issues found!'));
 
       final testResult = await expectSuccessfulProcessResult(
-        'flutter',
-        ['test', '--no-pub', '--coverage', '--reporter', 'compact'],
+        'dart',
+        ['test', '--coverage=coverage', '--reporter=compact'],
         workingDirectory: workingDirectory,
       );
       expect(testResult.stdout, contains('All tests passed!'));
+
+      // collect coverage
+      await expectSuccessfulProcessResult(
+        'dart',
+        [
+          'pub',
+          'run',
+          'coverage:format_coverage',
+          '--lcov',
+          '--in=coverage',
+          '--out=coverage/lcov.info',
+          '--packages=.dart_tool/package_config.json',
+          '--report-on=lib',
+        ],
+        workingDirectory: workingDirectory,
+      );
 
       final testCoverageResult = await expectSuccessfulProcessResult(
         'genhtml',
         ['coverage/lcov.info', '-o', 'coverage'],
         workingDirectory: workingDirectory,
       );
-      expect(testCoverageResult.stdout, contains('lines......: 97.8%'));
+      expect(testCoverageResult.stdout, contains('lines......: 100.0%'));
     }),
   );
 }
