+++ e2e/test/commands/create/dart_package/dart_pkg_test.dart
@@ -1,35 +1,24 @@
-@Tags(['e2e'])
-library legacy.dart_cli_test;
-
 import 'package:mason/mason.dart';
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
 import 'package:universal_io/io.dart';
 
-import '../../../../../helpers/helpers.dart';
+import '../../../../helpers/helpers.dart';
 
 void main() {
   test(
-    'create -t dart_cli',
+    'create dart_package',
     timeout: const Timeout(Duration(minutes: 2)),
     withRunner((commandRunner, logger, updater, logs) async {
       final tempDirectory = Directory.systemTemp.createTempSync();
       addTearDown(() => tempDirectory.deleteSync(recursive: true));
 
       final result = await commandRunner.run(
-        [
-          'create',
-          'very_good_dart_cli',
-          '-t',
-          'dart_cli',
-          '-o',
-          tempDirectory.path,
-        ],
+        ['create', 'dart_package', 'very_good_dart', '-o', tempDirectory.path],
       );
       expect(result, equals(ExitCode.success.code));
 
-      final workingDirectory =
-          path.join(tempDirectory.path, 'very_good_dart_cli');
+      final workingDirectory = path.join(tempDirectory.path, 'very_good_dart');
 
       // add coverage to collect coverage on dart test
       await expectSuccessfulProcessResult(
