+++ e2e/test/commands/create/legacy/dart_cli_test.dart
@@ -1,16 +1,13 @@
-@Tags(['e2e'])
-library flutter_pkg_test;
-
 import 'package:mason/mason.dart';
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
 import 'package:universal_io/io.dart';
 
-import '../../../../../helpers/helpers.dart';
+import '../../../../helpers/helpers.dart';
 
 void main() {
   test(
-    'create flutter_package',
+    'create -t dart_cli',
     timeout: const Timeout(Duration(minutes: 2)),
     withRunner((commandRunner, logger, updater, logs) async {
       final tempDirectory = Directory.systemTemp.createTempSync();
@@ -19,8 +16,9 @@ void main() {
       final result = await commandRunner.run(
         [
           'create',
-          'flutter_package',
-          'very_good_flutter',
+          'very_good_dart_cli',
+          '-t',
+          'dart_cli',
           '-o',
           tempDirectory.path,
         ],
@@ -28,7 +26,14 @@ void main() {
       expect(result, equals(ExitCode.success.code));
 
       final workingDirectory =
-          path.join(tempDirectory.path, 'very_good_flutter');
+          path.join(tempDirectory.path, 'very_good_dart_cli');
+
+      // add coverage to collect coverage on dart test
+      await expectSuccessfulProcessResult(
+        'dart',
+        ['pub', 'add', 'coverage:1.2.0'],
+        workingDirectory: workingDirectory,
+      );
 
       await expectSuccessfulProcessResult(
         'dart',
@@ -44,11 +49,27 @@ void main() {
       expect(analyzeResult.stdout, contains('No issues found!'));
 
       final testResult = await expectSuccessfulProcessResult(
-        'flutter',
-        ['test', '--no-pub', '--coverage', '--reporter', 'compact'],
+        'dart',
+        ['test', '--coverage=coverage', '--reporter=compact'],
         workingDirectory: workingDirectory,
       );
       expect(testResult.stdout, contains('All tests passed!'));
+
+      // collect coverage
+      await expectSuccessfulProcessResult(
+        'dart',
+        [
+          'pub',
+          'run',
+          'coverage:format_coverage',
+          '--lcov',
+          '--in=coverage',
+          '--out=coverage/lcov.info',
+          '--packages=.dart_tool/package_config.json',
+          '--report-on=lib',
+        ],
+        workingDirectory: workingDirectory,
+      );
 
       final testCoverageResult = await expectSuccessfulProcessResult(
         'genhtml',
