+++ e2e/test/commands/create/legacy/dart_pkg_test.dart
@@ -1,16 +1,13 @@
-@Tags(['e2e'])
-library dart_cli_test;
-
 import 'package:mason/mason.dart';
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
 import 'package:universal_io/io.dart';
 
-import '../../../../../helpers/helpers.dart';
+import '../../../../helpers/helpers.dart';
 
 void main() {
   test(
-    'create dart_cli',
+    'create -t dart_pkg',
     timeout: const Timeout(Duration(minutes: 2)),
     withRunner((commandRunner, logger, updater, logs) async {
       final tempDirectory = Directory.systemTemp.createTempSync();
@@ -19,16 +16,16 @@ void main() {
       final result = await commandRunner.run(
         [
           'create',
-          'dart_cli',
-          'very_good_dart_cli',
+          'very_good_dart',
+          '-t',
+          'dart_pkg',
           '-o',
-          tempDirectory.path,
+          tempDirectory.path
         ],
       );
       expect(result, equals(ExitCode.success.code));
 
-      final workingDirectory =
-          path.join(tempDirectory.path, 'very_good_dart_cli');
+      final workingDirectory = path.join(tempDirectory.path, 'very_good_dart');
 
       // add coverage to collect coverage on dart test
       await expectSuccessfulProcessResult(
