+++ e2e/test/commands/create/legacy/docs_site_test.dart
@@ -1,16 +1,13 @@
-@Tags(['e2e'])
-library docs_site_test;
-
 import 'package:mason/mason.dart';
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
 import 'package:universal_io/io.dart';
 
-import '../../../../../helpers/helpers.dart';
+import '../../../../helpers/helpers.dart';
 
 void main() {
   test(
-    'create docs_site',
+    'create -t docs_site',
     timeout: const Timeout(Duration(minutes: 2)),
     withRunner((commandRunner, logger, updater, logs) async {
       final tempDirectory = Directory.systemTemp.createTempSync();
@@ -19,8 +16,9 @@ void main() {
       final result = await commandRunner.run(
         [
           'create',
-          'docs_site',
           'very_good_docs_site',
+          '-t',
+          'docs_site',
           '-o',
           tempDirectory.path
         ],
