+++ e2e/test/commands/create/legacy/flame_game_test.dart
@@ -1,33 +1,32 @@
-@Tags(['e2e'])
-library dart_pkg_test;
-
 import 'package:mason/mason.dart';
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
 import 'package:universal_io/io.dart';
 
-import '../../../../../helpers/helpers.dart';
+import '../../../../helpers/helpers.dart';
 
 void main() {
   test(
-    'create dart_package',
+    'create -t flame_game',
     timeout: const Timeout(Duration(minutes: 2)),
     withRunner((commandRunner, logger, updater, logs) async {
       final tempDirectory = Directory.systemTemp.createTempSync();
-      addTearDown(() => tempDirectory.deleteSync(recursive: true));
 
       final result = await commandRunner.run(
-        ['create', 'dart_package', 'very_good_dart', '-o', tempDirectory.path],
+        [
+          'create',
+          'very_good_flame_game',
+          '-t',
+          'flame_game',
+          '-o',
+          tempDirectory.path,
+        ],
       );
       expect(result, equals(ExitCode.success.code));
 
-      final workingDirectory = path.join(tempDirectory.path, 'very_good_dart');
-
-      // add coverage to collect coverage on dart test
-      await expectSuccessfulProcessResult(
-        'dart',
-        ['pub', 'add', 'coverage:1.2.0'],
-        workingDirectory: workingDirectory,
+      final workingDirectory = path.join(
+        tempDirectory.path,
+        'very_good_flame_game',
       );
 
       await expectSuccessfulProcessResult(
@@ -44,34 +43,18 @@ void main() {
       expect(analyzeResult.stdout, contains('No issues found!'));
 
       final testResult = await expectSuccessfulProcessResult(
-        'dart',
-        ['test', '--coverage=coverage', '--reporter=compact'],
+        'flutter',
+        ['test', '--no-pub', '--coverage', '--reporter', 'compact'],
         workingDirectory: workingDirectory,
       );
       expect(testResult.stdout, contains('All tests passed!'));
-
-      // collect coverage
-      await expectSuccessfulProcessResult(
-        'dart',
-        [
-          'pub',
-          'run',
-          'coverage:format_coverage',
-          '--lcov',
-          '--in=coverage',
-          '--out=coverage/lcov.info',
-          '--packages=.dart_tool/package_config.json',
-          '--report-on=lib',
-        ],
-        workingDirectory: workingDirectory,
-      );
 
       final testCoverageResult = await expectSuccessfulProcessResult(
         'genhtml',
         ['coverage/lcov.info', '-o', 'coverage'],
         workingDirectory: workingDirectory,
       );
-      expect(testCoverageResult.stdout, contains('lines......: 100.0%'));
+      expect(testCoverageResult.stdout, contains('lines......: 97.8%'));
     }),
   );
 }
