+++ e2e/test/commands/create/legacy/flutter_pkg_test.dart
@@ -1,16 +1,13 @@
-@Tags(['e2e'])
-library legacy.docs_site_test;
-
 import 'package:mason/mason.dart';
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
 import 'package:universal_io/io.dart';
 
-import '../../../../../helpers/helpers.dart';
+import '../../../../helpers/helpers.dart';
 
 void main() {
   test(
-    'create -t docs_site',
+    'create -t flutter_pkg',
     timeout: const Timeout(Duration(minutes: 2)),
     withRunner((commandRunner, logger, updater, logs) async {
       final tempDirectory = Directory.systemTemp.createTempSync();
@@ -19,42 +16,44 @@ void main() {
       final result = await commandRunner.run(
         [
           'create',
-          'very_good_docs_site',
+          'very_good_flutter',
           '-t',
-          'docs_site',
+          'flutter_pkg',
           '-o',
-          tempDirectory.path
+          tempDirectory.path,
         ],
       );
       expect(result, equals(ExitCode.success.code));
 
       final workingDirectory =
-          path.join(tempDirectory.path, 'very_good_docs_site');
+          path.join(tempDirectory.path, 'very_good_flutter');
 
       await expectSuccessfulProcessResult(
-        'npm',
-        ['install'],
-        workingDirectory: workingDirectory,
-        validateStderr: false,
-      );
-
-      await expectSuccessfulProcessResult(
-        'npm',
-        ['run', 'format'],
+        'dart',
+        ['format', '--set-exit-if-changed', '.'],
         workingDirectory: workingDirectory,
       );
 
-      await expectSuccessfulProcessResult(
-        'npm',
-        ['run', 'lint'],
+      final analyzeResult = await expectSuccessfulProcessResult(
+        'flutter',
+        ['analyze', '.'],
         workingDirectory: workingDirectory,
       );
+      expect(analyzeResult.stdout, contains('No issues found!'));
 
-      await expectSuccessfulProcessResult(
-        'npm',
-        ['run', 'build'],
+      final testResult = await expectSuccessfulProcessResult(
+        'flutter',
+        ['test', '--no-pub', '--coverage', '--reporter', 'compact'],
         workingDirectory: workingDirectory,
       );
+      expect(testResult.stdout, contains('All tests passed!'));
+
+      final testCoverageResult = await expectSuccessfulProcessResult(
+        'genhtml',
+        ['coverage/lcov.info', '-o', 'coverage'],
+        workingDirectory: workingDirectory,
+      );
+      expect(testCoverageResult.stdout, contains('lines......: 100.0%'));
     }),
   );
 }
