+++ test/helpers/helpers.dart
@@ -1,3 +1,2 @@
 export 'command_helper.dart';
-export 'copy_directory.dart';
-export 'expect_successful_process_result.dart';
+export 'test_multi_template_commands.dart';
